+++ lib/wavefront-sdk/integration.rb
@@ -11,6 +11,7 @@ module Wavefront
     #
     # @param offset [Int] proxy at which the list begins
     # @param limit [Int] the number of proxies to return
+    # @return [Wavefront::Response]
     #
     def list(offset = 0, limit = 100)
       api.get('', offset: offset, limit: limit)
@@ -39,15 +40,15 @@ module Wavefront
       api.post([id, 'install'].uri_concat, nil)
     end
 
-    # POST /api/v2/integration/id/uninstall
-    # Uninstalls a Wavefront integration
+    # POST /api/v2/integration/{id}/install-all-alerts
+    # Enable all alerts associated with this integration
     #
     # @param id [String] ID of the integration
     # @return [Wavefront::Response]
     #
-    def uninstall(id)
+    def install_all_alerts(id)
       wf_integration_id?(id)
-      api.post([id, 'uninstall'].uri_concat, nil)
+      api.post([id, 'install-all-alerts'].uri_concat, nil)
     end
 
     # GET /api/v2/integration/id/status
@@ -61,13 +62,38 @@ module Wavefront
       api.get([id, 'status'].uri_concat)
     end
 
-    # GET /api/v2/integration/status
-    # Gets the status of all Wavefront integrations
+    # POST /api/v2/integration/id/uninstall
+    # Uninstalls a Wavefront integration
     #
+    # @param id [String] ID of the integration
     # @return [Wavefront::Response]
     #
-    def statuses
-      api.get('status')
+    def uninstall(id)
+      wf_integration_id?(id)
+      api.post([id, 'uninstall'].uri_concat, nil)
+    end
+
+    # POST /api/v2/integration/{id}/uninstall-all-alerts
+    # Disable all alerts associated with this integration
+    #
+    # @param id [String] ID of the integration
+    # @return [Wavefront::Response]
+    #
+    def uninstall_all_alerts(id)
+      wf_integration_id?(id)
+      api.post([id, 'uninstall-all-alerts'].uri_concat, nil)
+    end
+
+    # GET /api/v2/integration/installed
+    # Gets a flat list of all Integrations that are installed, along
+    # with their status
+    #
+    # @param offset [Int] proxy at which the list begins
+    # @param limit [Int] the number of proxies to return
+    # @return [Wavefront::Response]
+    #
+    def installed
+      api.get('installed')
     end
 
     # GET /api/v2/integration/manifests
@@ -76,6 +102,15 @@ module Wavefront
     #
     def manifests
       api.get('manifests')
+    end
+
+    # GET /api/v2/integration/status
+    # Gets the status of all Wavefront integrations
+    #
+    # @return [Wavefront::Response]
+    #
+    def statuses
+      api.get('status')
     end
   end
 end